PM-3323: Fix the merge.
